refactor: move bridge handling logic to client
