+++ src/lib/proof.ts
@@ -1,6 +1,5 @@
-import type { BridgeInitialData } from "@/pages/api/pair-client";
 import verificationKeys from "@/public/semaphore/verification_key.json";
-import type { FP, Verification } from "@/types";
+import type { BridgeInitialData, FP, Verification } from "@/types";
 import { ProofError, type Identity } from "@/types";
 import { Group } from "@semaphore-protocol/group";
 import { Identity as ZkIdentity } from "@semaphore-protocol/identity";
