+++ src/services/bridge/index.ts
@@ -1,0 +1,2 @@
+export * from "./approve-request";
+export * from "./pair-client";
