+++ src/types/bridge-request-data.ts
@@ -1,0 +1,4 @@
+export type BridgeRequestData = {
+  iv: string;
+  payload: string;
+};
