+++ src/types/bridge-service-return-type.ts
@@ -1,0 +1,10 @@
+export type BridgeServiceReturnType<
+  SuccessPayload extends Record<string, unknown> = NonNullable<unknown>,
+> =
+  | {
+      success: false;
+      error: Error;
+    }
+  | (SuccessPayload & {
+      success: true;
+    });
